requirements.txt
flask==2.3.2
cx_Oracle==8.3.0
python-dotenv==1.0.0